# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "aifn.run API",
    "version": "2.0.0",
    "description": "Versioned AI functions with predictable text or JSON output. The unversioned function URL always resolves to the latest version."
  },
  "servers": [
    {
      "url": "https://aifn.run"
    }
  ],
  "paths": {
    "/auth/login": {
      "get": {
        "operationId": "startLogin",
        "summary": "Start the OIDC authorization-code flow",
        "parameters": [{"name": "return_to", "in": "query", "schema": {"type": "string"}}],
        "responses": {"302": {"description": "Redirect to the configured OIDC provider"}}
      }
    },
    "/auth/callback": {
      "get": {
        "operationId": "completeLogin",
        "summary": "Complete the OIDC authorization-code flow",
        "responses": {"302": {"description": "Redirect to the requested application page"}, "400": {"description": "Authentication failed"}}
      }
    },
...
  },
...
}