# server/index.ts
import { createServer, type IncomingMessage, type ServerResponse } from 'node:http';
import { readFile, stat } from 'node:fs/promises';
import { createReadStream } from 'node:fs';
import { extname, join, normalize } from 'node:path';
import { fileURLToPath } from 'node:url';
import { randomUUID } from 'node:crypto';
import { createAuth } from './auth.js';
import migrationInitialSchema from './migrations/001-initial-schema.js';
import migrationProviderEndpoint from './migrations/002-provider-endpoint.js';
import migrationFunctionVisibility from './migrations/003-function-visibility.js';
import migrationOidcSessions from './migrations/004-oidc-sessions.js';
import type { Migration } from './migrations/types.js';

type OutputMode = 'json' | 'text';
type Database = {
  get(sql: string, data?: unknown): Promise<any>;
  run(sql: string, data?: unknown): Promise<any>;
  all(sql: string, data?: unknown): Promise<any[]>;
  pragma?: (values: string[]) => void;
};
type FunctionVersion = {
  functionId: string;
  version: number;
  prompt: string;
  name: string;
  model: string;
  format: string;
  output: OutputMode;
  providerEndpoint: string;
  public: boolean;
  active?: boolean;
};

const root = fileURLToPath(new URL('..', import.meta.url));
const port = Number(process.env.PORT || 3000);
const cacheControl = 'public, max-age=604800, must-revalidate';
const revalidateControl = 'public, max-age=86400, must-revalidate';
const uuid = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
const openapiJSON = await readFile(join(root, 'openapi/openapi.json'), 'utf8');
const openapiYAML = await readFile(join(root, 'openapi/openapi.yaml'), 'utf8');
const database = await loadDatabase();
database.pragma?.(['foreign_keys = ON']);
await migrate();
const auth = createAuth(database);

async function loadDatabase(): Promise<Database> {
  const address = process.env.DATABASE_URL;
  if (!address) {
    throw new Error('DATABASE_URL is required');
  }

  const url = new URL(address);
  const module = await import(url.href);
  const db = module.default || module;
  return db;
}

async function migrate() {
  await database.run(`CREATE TABLE IF NOT EXISTS schema_migrations (id TEXT PRIMARY KEY, applied_at TEXT NOT NULL)`);
  const applied = new Set((await database.all(`SELECT id FROM schema_migrations`)).map((row) => row.id));
  const migrations: Migration[] = [migrationInitialSchema, migrationProviderEndpoint, migrationFunctionVisibility, migrationOidcSessions];
  for (const migration of migrations) {
    if (applied.has(migration.id)) continue;
    await migration.up(database);
    await database.run(`INSERT INTO schema_migrations (id, applied_at) VALUES (?, ?)`, [migration.id, new Date().toISOString()]);
  }
}

function send(
  res: ServerResponse,
  status: number,
  body: unknown,
  type = 'application/json; charset=utf-8',
  headers: Record<string, string> = {},
) {
  const content = typeof body === 'string' ? body : JSON.stringify(body);
  res.writeHead(status, { 'content-type': type, 'content-length': Buffer.byteLength(content), ...headers });
  res.end(content);
  return true;
}

function error(res: ServerResponse, status: number, code: string, message: string) {
  return send(res, status, { error: { code, message } });
}
function redirect(res: ServerResponse, location: string, headers: Record<string, string> = {}) {
  res.writeHead(302, { location, ...headers });
  res.end();
  return true;
}
function body(req: IncomingMessage) {
  return new Promise<string>((resolve, reject) => {
    const chunks: Buffer[] = [];
    req.on('data', (chunk) => chunks.push(chunk));
    req.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')));
    req.on('error', reject);
  });
}
function requestOf(req: IncomingMessage) {
  return new Request(`http://${req.headers.host || 'localhost'}${req.url || '/'}`, {
    headers: { cookie: req.headers.cookie || '' },
  });
}
async function profileId(req: IncomingMessage) {
  return (await auth.session(requestOf(req)))?.id || null;
}
async function requireAuthentication(req: IncomingMessage, res: ServerResponse) {
  if (await auth.session(requestOf(req))) return true;
  error(res, 401, 'AUTHENTICATION_REQUIRED', 'Authentication required');
  return false;
}
function versionJSON(row: any): FunctionVersion {
  return {
    functionId: row.function_id,
    version: Number(row.version),
    prompt: row.prompt,
    name: row.name,
    model: row.model,
    format: row.format,
    output: row.output,
    providerEndpoint: row.provider_endpoint || '',
    public: Boolean(row.is_public),
    active: Boolean(row.active),
  };
}

async function getVersion(id: string, version?: number) {
  const row = version
    ? await database.get(
        `SELECT v.*, f.is_public, f.active_version = v.version AS active FROM function_versions v JOIN functions f ON f.id = v.function_id WHERE v.function_id = ? AND v.version = ?`,
        [id, version],
      )
    : await database.get(
        `SELECT v.*, f.is_public, 1 AS active FROM function_versions v JOIN functions f ON f.id = v.function_id WHERE v.function_id = f.id AND v.version = f.active_version`,
      );
  return row ? versionJSON(row) : null;
}

async function saveFunction(req: IncomingMessage, res: ServerResponse, id?: string) {
  if (!(await requireAuthentication(req, res))) return true;
  let input: any;
  try {
    input = JSON.parse(await body(req));
  } catch {
    return error(res, 400, 'INVALID_JSON', 'Request body must be JSON');
  }
  if (!input.prompt || typeof input.prompt !== 'string') return error(res, 400, 'INVALID_PROMPT', 'prompt is required');
  if (input.output && !['json', 'text'].includes(input.output))
    return error(res, 400, 'INVALID_OUTPUT', 'output must be json or text');
  if (input.public !== undefined && typeof input.public !== 'boolean')
    return error(res, 400, 'INVALID_VISIBILITY', 'public must be a boolean');
  const functionId = id || randomUUID();
  const now = new Date().toISOString();
  const current = id ? await database.get(`SELECT * FROM functions WHERE id = ?`, [id]) : null;
  if (id && !current) return error(res, 404, 'FUNCTION_NOT_FOUND', 'Function not found');
  const version = current ? Number(current.latest_version) + 1 : 1;
  if (!current)
    await database.run(
      `INSERT INTO functions (id, owner_id, is_public, active_version, latest_version, created_at, updated_at) VALUES (?, ?, ?, 1, 1, ?, ?)`,
      [functionId, null, input.public === true ? 1 : 0, now, now],
    );
  else
    await database.run(`UPDATE functions SET is_public = ?, latest_version = ?, active_version = ?, updated_at = ? WHERE id = ?`, [
      input.public === undefined ? Number(current.is_public) : input.public === true ? 1 : 0,
      version,
      version,
      now,
      functionId,
    ]);
  await database.run(
    `INSERT INTO function_versions (function_id, version, prompt, name, model, format, output, provider_endpoint, created_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)`,
    [
      functionId,
      version,
      input.prompt,
      input.name || '',
      input.model || '',
      input.format || 'chat',
      input.output || 'json',
      input.providerEndpoint || '',
      now,
    ],
  );
  if (typeof input.providerKey === 'string' && input.providerKey)
    await auth.setProperty(requestOf(req), `aifn:provider:function:${functionId}:key`, input.providerKey);
  return send(res, 201, await getVersion(functionId, version));
}

function interpolate(prompt: string, input: unknown) {
  if (typeof input === 'string') return `${prompt}\n${input}`;
  return prompt.replace(/\{([\s\S]+?)\}/g, (_, key) => String((input as Record<string, unknown>)?.[key.trim()] ?? ''));
}
async function complete(fn: FunctionVersion, input: unknown, req: IncomingMessage) {
  const content = interpolate(fn.prompt, input);
  const json = fn.output === 'json';
  const accountEndpoint = (await auth.property(requestOf(req), 'aifn:provider:account:endpoint'))?.value || '';
  const accountKey = (await auth.property(requestOf(req), 'aifn:provider:account:key'))?.value || '';
  const functionKey = (await auth.property(requestOf(req), `aifn:provider:function:${fn.functionId}:key`))?.value || '';
  const endpoint = fn.providerEndpoint || accountEndpoint || process.env.API_CHAT_URL || '';
  const key = functionKey || accountKey || process.env.API_KEY || '';
  const messages = [
    {
      role: 'system',
      content: json
        ? 'Return only valid JSON. Do not use Markdown fences or commentary.'
        : process.env.SYSTEM_MESSAGE || '',
    },
    { role: 'user', content },
  ].filter((message) => message.content);
  const response = await fetch(endpoint, {
    method: 'POST',
    headers: { 'content-type': 'application/json', authorization: `Bearer ${key}` },
    body: JSON.stringify({
      model: fn.model || process.env.API_MODEL,
      messages,
      response_format: json ? { type: 'json_object' } : undefined,
    }),
  });
  if (!response.ok) throw new Error(`Provider returned ${response.status}`);
  const result: any = await response.json();
  const text = result.choices?.map((choice: any) => choice.message?.content || '').join('\n') || '';
  if (!json) return text;
  try {
    return JSON.parse(text.replace(/^```json\s*|\s*```$/g, ''));
  } catch {
    throw new Error('Provider returned invalid JSON');
  }
}

async function requireFunctionAccess(req: IncomingMessage, res: ServerResponse, fn: FunctionVersion) {
  if (fn.public || (await auth.session(requestOf(req)))) return true;
  error(res, 404, 'FUNCTION_NOT_FOUND', 'Function not found');
  return false;
}

async function api(req: IncomingMessage, res: ServerResponse, url: URL) {
  if (url.pathname === '/api') {
    const yaml = (req.headers.accept || '').includes('yaml') || url.searchParams.get('format') === 'yaml';
    return send(
      res,
      200,
      yaml ? openapiYAML : openapiJSON,
      yaml ? 'application/yaml; charset=utf-8' : 'application/json; charset=utf-8',
      { 'cache-control': cacheControl },
    );
  }
  const parts = url.pathname.split('/').filter(Boolean);
  if (parts[0] !== 'api') return false;
  if (parts[1] === 'provider' && parts[2] === 'account') {
    if (!(await auth.session(requestOf(req))))
      return error(res, 401, 'AUTHENTICATION_REQUIRED', 'Authentication required');
    if (req.method === 'GET') {
      const endpoint = await auth.property(requestOf(req), 'aifn:provider:account:endpoint');
      const model = await auth.property(requestOf(req), 'aifn:provider:account:model');
      return send(res, 200, {
        endpoint: endpoint?.value || '',
        model: model?.value || '',
        keyConfigured: Boolean((await auth.property(requestOf(req), 'aifn:provider:account:key'))?.value),
      });
    }
    if (req.method === 'PUT') {
      let input: any;
      try {
        input = JSON.parse(await body(req));
      } catch {
        return error(res, 400, 'INVALID_JSON', 'Request body must be JSON');
      }
      try {
        if (input.endpoint !== undefined)
          await auth.setProperty(requestOf(req), 'aifn:provider:account:endpoint', String(input.endpoint));
        if (input.model !== undefined)
          await auth.setProperty(requestOf(req), 'aifn:provider:account:model', String(input.model));
        if (input.key !== undefined && input.key !== '')
          await auth.setProperty(requestOf(req), 'aifn:provider:account:key', String(input.key));
        return send(res, 200, {
          endpoint: input.endpoint || '',
          model: input.model || '',
          keyConfigured: Boolean(input.key),
        });
      } catch (cause) {
        return error(
          res,
          502,
          'AUTH_PROPERTY_ERROR',
          cause instanceof Error ? cause.message : 'Unable to save provider settings',
        );
      }
    }
  }
  if (parts[1] === 'fn' && parts.length === 2 && req.method === 'GET' && url.searchParams.get('visibility') === 'all') {
    if (!(await requireAuthentication(req, res))) return true;
    const rows = await database.all(
      `SELECT f.id AS function_id, f.active_version AS version, v.name, v.output, f.is_public FROM functions f JOIN function_versions v ON v.function_id = f.id AND v.version = f.active_version`,
    );
    return send(res, 200, rows.map((row) => ({ functionId: row.function_id, version: Number(row.version), name: row.name, output: row.output, public: Boolean(row.is_public) })));
  }
  if (parts[1] === 'fn' && parts.length === 2 && req.method === 'GET') {
    const rows = await database.all(
      `SELECT f.id AS function_id, f.active_version AS version, v.name, v.output, f.is_public FROM functions f JOIN function_versions v ON v.function_id = f.id AND v.version = f.active_version WHERE f.is_public = 1`,
    );
    return send(
      res,
      200,
      rows.map((row) => ({
        functionId: row.function_id,
        version: Number(row.version),
        name: row.name,
        output: row.output,
        public: Boolean(row.is_public),
      })),
    );
  }
  if (parts[1] === 'fn' && parts.length === 2 && req.method === 'POST') return saveFunction(req, res);
  if (parts[1] === 'fn' && uuid.test(parts[2] || '')) {
    const id = parts[2];
    const versionValue = parts[3] ? Number(parts[3]) : undefined;
    const invalidVersion = parts[3] !== undefined && (!Number.isInteger(versionValue) || (versionValue ?? 0) < 1);
    if (parts.length > 4 || invalidVersion) return error(res, 400, 'INVALID_VERSION', 'Invalid function version');
    if (req.method === 'GET') {
      const found = await getVersion(id, versionValue);
      if (!found) return error(res, 404, 'VERSION_NOT_FOUND', 'Function version not found');
      if (!(await requireFunctionAccess(req, res, found))) return true;
      return send(res, 200, found);
    }
    if (req.method === 'PUT' && versionValue === undefined) return saveFunction(req, res, id);
    if (req.method === 'DELETE' && versionValue === undefined) {
      if (!(await requireAuthentication(req, res))) return true;
      const result = await database.run(`DELETE FROM functions WHERE id = ?`, [id]);
      return result ? res.writeHead(204).end() : error(res, 404, 'FUNCTION_NOT_FOUND', 'Function not found');
    }
  }
  if (parts[1] === 'run' && uuid.test(parts[2] || '') && req.method === 'POST') {
    const fn = await getVersion(
      parts[2],
      url.searchParams.get('version') ? Number(url.searchParams.get('version')) : undefined,
    );
    if (!fn) return error(res, 404, 'VERSION_NOT_FOUND', 'Function version not found');
    if (!(await requireFunctionAccess(req, res, fn))) return true;
    try {
      const raw = await body(req);
      let input: unknown = raw;
      try {
        input = JSON.parse(raw).inputs ?? {};
      } catch {}
      const output = await complete(fn, input, req);
      return send(
        res,
        200,
        output,
        fn.output === 'json' ? 'application/json; charset=utf-8' : 'text/plain; charset=utf-8',
        { 'x-aifn-function-id': fn.functionId, 'x-aifn-version': String(fn.version), 'cache-control': 'no-store' },
      );
    } catch (cause) {
      return error(res, 502, 'PROVIDER_ERROR', cause instanceof Error ? cause.message : 'Provider request failed');
    }
  }
  return false;
}

async function staticFile(_req: IncomingMessage, res: ServerResponse, url: URL) {
  const dist = join(root, 'dist');
  const requested = url.pathname === '/' ? '/index.html' : url.pathname;
  const candidate = normalize(join(dist, requested));
  if (!candidate.startsWith(dist)) return error(res, 403, 'FORBIDDEN', 'Forbidden');
  const file = extname(candidate) ? candidate : join(dist, 'index.html');
  try {
    const info = await stat(file);
    if (!info.isFile()) return false;
    const mime: Record<string, string> = {
      '.html': 'text/html; charset=utf-8',
      '.js': 'text/javascript; charset=utf-8',
      '.css': 'text/css; charset=utf-8',
      '.webmanifest': 'application/manifest+json; charset=utf-8',
      '.svg': 'image/svg+xml',
    };
    const extension = extname(file);
    res.writeHead(200, {
      'content-type': mime[extension] || 'application/octet-stream',
      'cache-control': revalidateControl,
    });
    createReadStream(file).pipe(res);
    return true;
  } catch {
    return false;
  }
}

const server = createServer(async (req: IncomingMessage, res: ServerResponse) => {
  try {
    const url = new URL(req.url || '/', `http://${req.headers.host || 'localhost'}`);
    const protocol = req.headers['x-forwarded-proto']?.toString().split(',')[0] || url.protocol.replace(':', '');
    const origin = `${protocol}://${req.headers.host || url.host}`;
    const redirectUri = `${origin}/auth/callback`;
    if (url.pathname === '/auth/login' && req.method === 'GET') return redirect(res, await auth.login(url.searchParams.get('return_to') || '/dashboard', redirectUri));
    if (url.pathname === '/auth/callback' && req.method === 'GET') {
      try {
        const result = await auth.callback(url.searchParams.get('code') || '', url.searchParams.get('state') || '', redirectUri);
        return redirect(res, result.returnTo, { 'set-cookie': `${auth.sessionCookie}=${result.id}; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=2592000` });
      } catch (cause) {
        return error(res, 400, 'AUTHENTICATION_FAILED', cause instanceof Error ? cause.message : 'Authentication failed');
      }
    }
    if (url.pathname === '/auth/session' && req.method === 'GET') return send(res, 200, { profile: await auth.session(req) });
    if (url.pathname === '/auth/logout' && (req.method === 'POST' || req.method === 'GET')) {
      await auth.logout(req);
      return redirect(res, '/', { 'set-cookie': `${auth.sessionCookie}=; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=0` });
    }
    if (req.method === 'OPTIONS') {
      res
        .writeHead(204, {
          'access-control-allow-origin': '*',
          'access-control-allow-methods': 'GET,POST,PUT,DELETE,OPTIONS',
          'access-control-allow-headers': 'content-type,authorization',
        })
        .end();
      return;
    }
    if (url.pathname.startsWith('/api')) {
      if (await api(req, res, url)) return;
      return error(res, 404, 'NOT_FOUND', 'API route not found');
    }
    if (await staticFile(req, res, url)) return;
    error(res, 404, 'NOT_FOUND', 'Not found');
  } catch (cause) {
    error(res, 500, 'INTERNAL_ERROR', cause instanceof Error ? cause.message : 'Internal server error');
  }
});
server.listen(port, '0.0.0.0', () => console.log(`aifn.run listening on ${port}`));